+++ src/eyas-interface/app/src/App.vue
@@ -8,6 +8,7 @@
 		<EnvironmentModal />
 		<VariablesModal />
 		<ExitModal />
+		<VersionMismatchModal />
 	</v-app>
 </template>
 
@@ -15,12 +16,14 @@
 import ExitModal from '@/components/ExitModal.vue';
 import EnvironmentModal from '@/components/EnvironmentModal.vue';
 import VariablesModal from '@/components/VariablesModal.vue';
+import VersionMismatchModal from '@/components/VersionMismatchModal.vue';
 
 export default {
 	components: {
 		EnvironmentModal,
 		VariablesModal,
-		ExitModal
+		ExitModal,
+		VersionMismatchModal
 	}
 };
 </script>